+++ plotlykt-core/src/jsMain/kotlin/space/kscience/plotly/PlotlyElement.kt
@@ -4,14 +4,20 @@ import kotlinx.html.TagConsumer
 import kotlinx.html.div
 import kotlinx.serialization.ExperimentalSerializationApi
 import kotlinx.serialization.json.Json
+import kotlinx.serialization.json.decodeFromDynamic
 import kotlinx.serialization.json.encodeToDynamic
 import org.w3c.dom.Element
 import org.w3c.dom.HTMLElement
+import org.w3c.dom.events.MouseEvent
 import space.kscience.dataforge.meta.MetaSerializer
 import space.kscience.dataforge.meta.Scheme
+import space.kscience.dataforge.meta.Value
 import space.kscience.dataforge.names.asName
 import space.kscience.dataforge.names.firstOrNull
 import space.kscience.dataforge.names.startsWith
+import space.kscience.plotly.events.PlotlyEvent
+import space.kscience.plotly.events.PlotlyEventListenerType
+import space.kscience.plotly.events.PlotlyEventPoint
 
 @OptIn(ExperimentalSerializationApi::class)
 private fun Scheme.toDynamic(): dynamic = Json.encodeToDynamic(MetaSerializer, meta)
@@ -19,7 +25,7 @@ private fun Scheme.toDynamic(): dynamic = Json.encodeToDynamic(MetaSerializer, m
 private fun List<Scheme>.toDynamic(): Array<dynamic> = map { it.toDynamic() }.toTypedArray()
 
 /**
- * Attach a plot to this element or update existing plot
+ * Attach a plot to this element or update the existing plot
  */
 public fun Element.plot(plotlyConfig: PlotlyConfig = PlotlyConfig(), plot: Plot) {
     val tracesData = plot.data.toDynamic()
@@ -58,19 +64,43 @@ public fun Element.plot(plotlyConfig: PlotlyConfig = PlotlyConfig(), plot: Plot)
 @Deprecated("Change arguments positions", ReplaceWith("plot(plotlyConfig, plot)"))
 public fun Element.plot(plot: Plot, plotlyConfig: PlotlyConfig = PlotlyConfig()): Unit = plot(plotlyConfig, plot)
 
+/**
+ * Create a plot in this element
+ */
 public inline fun Element.plot(plotlyConfig: PlotlyConfig = PlotlyConfig(), plotBuilder: Plot.() -> Unit) {
     plot(plotlyConfig, Plot().apply(plotBuilder))
 }
 
+public class PlotlyElement(public val div: HTMLElement)
+
+/**
+ * Create a div element and render plot in it
+ */
 public fun TagConsumer<HTMLElement>.plotDiv(
     plotlyConfig: PlotlyConfig = PlotlyConfig(),
     plot: Plot,
-): HTMLElement = div("plotly-kt-plot").apply { plot(plotlyConfig, plot) }
+): PlotlyElement = PlotlyElement(div("plotly-kt-plot").apply { plot(plotlyConfig, plot) })
 
 /**
- * Render plot in HTML element using direct plotly API.
+ * Render plot in the HTML element using direct plotly API.
  */
 public inline fun TagConsumer<HTMLElement>.plotDiv(
     plotlyConfig: PlotlyConfig = PlotlyConfig(),
     plotBuilder: Plot.() -> Unit,
-): HTMLElement = div("plotly-kt-plot").apply { plot(plotlyConfig, plotBuilder) }
+): PlotlyElement = PlotlyElement(div("plotly-kt-plot").apply { plot(plotlyConfig, plotBuilder) })
+
+@OptIn(ExperimentalSerializationApi::class)
+public fun PlotlyElement.on(eventType: PlotlyEventListenerType, block: MouseEvent.(PlotlyEvent) -> Unit) {
+    div.asDynamic().on(eventType.eventType) { event: PlotMouseEvent ->
+        val eventData = PlotlyEvent(event.points.map {
+            PlotlyEventPoint(
+                curveNumber = it.curveNumber as Int,
+                pointNumber = it.pointNumber as? Int,
+                x = Value.of(it.x),
+                y = Value.of(it.y),
+                data = Json.decodeFromDynamic(it.data)
+            )
+        })
+        event.event.block(eventData)
+    }
+}