+++ plotlykt-core/src/jvmTest/kotlin/space/kscience/plotly/PlotSerializationTest.kt
@@ -1,11 +1,15 @@
 package space.kscience.plotly
 
+import kotlinx.serialization.json.JsonArray
+import kotlinx.serialization.json.jsonObject
 import org.junit.jupiter.api.Test
+import space.kscience.dataforge.meta.ListValue
 import space.kscience.dataforge.meta.MutableMeta
 import space.kscience.dataforge.meta.asObservable
-import space.kscience.dataforge.meta.ListValue
+import space.kscience.plotly.models.ShapeType
 import space.kscience.plotly.models.TraceType
 import kotlin.test.assertEquals
+import kotlin.test.assertTrue
 
 
 class PlotSerializationTest {
@@ -23,5 +27,19 @@ class PlotSerializationTest {
         assertEquals(1, plot.data.size)
         assertEquals(TraceType.scatter, plot.data[0].type)
         assertEquals(1.0, plot.data[0].x.doubles[0])
+    }
+
+    @Test
+    fun shapeSerialization(){
+        val plot = Plotly.plot {
+            shape {
+                type = ShapeType.rect
+            }
+        }
+
+        val json = plot.toJson()
+        println(json.toString())
+        val shapes = json["layout"]?.jsonObject?.get("shapes")
+        assertTrue { shapes is JsonArray }
     }
 }